GP2-621 update objective, refactor file structure
