+++ exportplan/templates/exportplan/sections/objectives.html
@@ -30,9 +30,9 @@
       </div>
     </div>
     <div class="c-1-2 text-blue-deep-80">
-      <h2 class="h-l">What you want to achieve</h2>
+      <h2 class="h-m">What you want to achieve</h2>
       <p>What do you want to accomplish and by when?</p>
-      <div class="g-panel">
+      <div class="g-panel m-b-m">
         <p>To increase the prospect of success, ensure what you want to achieve is specific, measurable and realistic:</p>
         <ul class="list-dot">
           <li>specific - your objective is well defined</li>
