--- react-components/src/views/sections/Objectives/ObjectivesReasons/index.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-import { FormWithInputWithExample } from '@src/components/FormWithInputWithExample'
-
-export const ObjectivesReasons = (params) => (
-  <FormWithInputWithExample {...params} />
-)
